fix(tests): consolidate ENV_MUTEX into shared test_support module
Previous per-module mutex still let tests in DIFFERENT submodules
race (policy::tests vs policy::env::tests vs policy::ffi::tests)
because each had its own static ENV_MUTEX.  Tests in module A could
overwrite env vars while a test in module B was mid-configure_*.

Move the mutex to `policy::test_support::ENV_MUTEX` (single
process-wide static) and have each test module's EnvGuard delegate
to a single `EnvGuard::new(&[keys])` constructor.  All env-touching
tests across the three submodules now serialize correctly.
